Refactored loading Koenig-Lexical package
- DRY up all the fetching so we only do it from one place
