+++ apps/admin-x-settings/src/admin-x-ds/global/form/HtmlField.tsx
@@ -4,13 +4,7 @@ import HtmlEditor, {HtmlEditorProps} from './HtmlEditor';
 import React from 'react';
 import clsx from 'clsx';
 
-export type EditorConfig = { editor: { url: string; version: string; } }
-
 export type HtmlFieldProps = HtmlEditorProps & {
-    /**
-     * Should be passed the Ghost instance config to get the editor JS URL
-     */
-    config: EditorConfig;
     title?: string;
     hideTitle?: boolean;
     error?: boolean;
