+++ apps/admin-x-settings/src/admin-x-ds/providers/DesignSystemProvider.tsx
@@ -4,6 +4,7 @@ import React, {createContext, useContext, useState} from 'react';
 interface DesignSystemContextType {
     isAnyTextFieldFocused: boolean;
     setFocusState: (value: boolean) => void;
+    fetchKoenigLexical: any;
 }
 
 const DesignSystemContext = createContext<DesignSystemContextType | undefined>(undefined);
@@ -17,10 +18,11 @@ export const useFocusContext = () => {
 };
 
 interface DesignSystemProviderProps {
+    fetchKoenigLexical: any;
     children: React.ReactNode;
 }
 
-const DesignSystemProvider: React.FC<DesignSystemProviderProps> = ({children}) => {
+const DesignSystemProvider: React.FC<DesignSystemProviderProps> = ({fetchKoenigLexical, children}) => {
     const [isAnyTextFieldFocused, setIsAnyTextFieldFocused] = useState(false);
 
     const setFocusState = (value: boolean) => {
@@ -28,7 +30,7 @@ const DesignSystemProvider: React.FC<DesignSystemProviderProps> = ({children}) =
     };
 
     return (
-        <DesignSystemContext.Provider value={{isAnyTextFieldFocused, setFocusState}}>
+        <DesignSystemContext.Provider value={{isAnyTextFieldFocused, fetchKoenigLexical, setFocusState}}>
             {children}
         </DesignSystemContext.Provider>
     );
