+++ apps/admin-x-settings/src/main.tsx
@@ -8,6 +8,7 @@ ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
     <React.StrictMode>
         <App
             externalNavigate={() => {}}
+            fetchKoenigLexical={() => {}}
             ghostVersion='5.x'
             officialThemes={[{
                 name: 'Casper',
